FittedItemBanks.jl

This module provides abstract and concrete item banks, which store information about items and their parameters such as difficulty, most typically resulting from fitting an Item-Response Theory (IRT) model.

FittedItemBanks.FittedItemBanks

FittedItemBanks.CdfMirtItemBank

FittedItemBanks.PerCategoryFloat

FittedItemBanks.ItemBank2PL

FittedItemBanks.ItemBank3PL

FittedItemBanks.ItemBank4PL

FittedItemBanks.ItemBankMirt2PL

FittedItemBanks.ItemBankMirt3PL

FittedItemBanks.ItemBankMirt4PL

FittedItemBanks._search

FittedItemBanks.item_bank_domain

FittedItemBanks.FittedItemBanks — Module

This module provides abstract and concrete item banks, which store information about items and their parameters such as difficulty, most typically resulting from fitting an Item-Response Theory (IRT) model.

FittedItemBanks.CdfMirtItemBank — Type

This item bank corresponds to the most commonly found version of MIRT in the literature. Its items feature multidimensional discriminations and its learners multidimensional abilities, but item difficulties are single-dimensional.

FittedItemBanks.PerCategoryFloat — Type

This item bank implements the nominal model. The Graded Partial Credit Model (GPCM) is implemented in terms of this one. See:

A Generalized Partial Credit Model: Application of an EM Algorithm Muraki, E., (1992). Applied Psychological Measurement 10.1177/014662169201600206

And/or

A Generalized Partial Credit Model Muraki, E. (1997). In Handbook of Modern Item Response Theory. Springer, New York, NY. https://doi.org/10.1007/978-1-4757-2691-6_9

Currently, this item bank only supports the normal scaled logistic as the characteristic/transfer function.

FittedItemBanks.ItemBank2PL — Method

Convenience function to construct an item bank of the standard 2-parameter logistic single-dimensional IRT model.

FittedItemBanks.ItemBank3PL — Method

Convenience function to construct an item bank of the standard 3-parameter logistic single-dimensional IRT model.

FittedItemBanks.ItemBank4PL — Method

Convenience function to construct an item bank of the standard 4-parameter logistic single-dimensional IRT model.

FittedItemBanks.ItemBankMirt2PL — Method

Convenience function to construct an item bank of the standard 2-parameter logistic MIRT model.

FittedItemBanks.ItemBankMirt3PL — Method

Convenience function to construct an item bank of the standard 3-parameter logistic MIRT model.

FittedItemBanks.ItemBankMirt4PL — Method

Convenience function to construct an item bank of the standard 4-parameter logistic MIRT model.

FittedItemBanks._search — Method

Binary search for the point x where f(x) = target += precis given f is assumed as monotonically increasing.

FittedItemBanks.item_bank_domain — Method

Given an item bank, this function returns the domain of the item bank, i.e. the range (lo, hi) which includes for each item the range in which the the item response function is changing.