Generic interface

This page details the operations which should be supported by different ItemResponse elements, as well as traits for categorisation which can be used to dispatch to different operations.

Basic types

FittedItemBanks.AbstractItemBank — Type

abstract type AbstractItemBank

Base supertype for all item banks.

FittedItemBanks.ItemResponse — Type

struct ItemResponse{ItemBankT<:AbstractItemBank}

item_bank::AbstractItemBank

index::Int64

An item response.

AbstractItemBank traits

Domain

FittedItemBanks.DomainType — Type

abstract type DomainType

DomainType(::AbstractItemBank) -> DomainType

Domain type for a item banks' item response functions. Used as a trait.

FittedItemBanks.DiscreteDomain — Type

abstract type DiscreteDomain <: DomainType

A discrete domain. Typically this is a sampled version of a continuous domain item bank.

Item response functions with discrete domains tend to support less operations than those with continuous domains.

FittedItemBanks.ContinuousDomain — Type

abstract type ContinuousDomain <: DomainType

A continuous domain.

FittedItemBanks.VectorContinuousDomain — Type

struct VectorContinuousDomain <: ContinuousDomain

A continuous domain that is vector valued.

FittedItemBanks.OneDimContinuousDomain — Type

struct OneDimContinuousDomain <: ContinuousDomain

A continuous domain that is scalar valued.

FittedItemBanks.DiscreteIndexableDomain — Type

struct DiscreteIndexableDomain <: DiscreteDomain

An discrete domain which is efficiently indexable and iterable.

FittedItemBanks.DiscreteIterableDomain — Type

struct DiscreteIterableDomain <: DiscreteDomain

An discrete domain which is only efficiently iterable.

Response

FittedItemBanks.ResponseType — Type

abstract type ResponseType

ResponseType(::AbstractItemBank) -> ResponseType

A response type for an item bank. Used as a trait.

FittedItemBanks.BooleanResponse — Type

struct BooleanResponse <: ResponseType

A boolean/dichotomous response.

FittedItemBanks.MultinomialResponse — Type

struct MultinomialResponse <: ResponseType

A multinomial response, including ordinal responses.

AbstractItemBank methods

Base.length — Method

length(item_bank::AbstractItemBank)

Returns the number of items in the item bank.

FittedItemBanks.subset — Function

function subset(item_bank::AbstractItemBank, idxs)

Return a new item bank of the same type, with the items at the given indices.

FittedItemBanks.item_bank_domain — Function

item_bank_domain(
    item_bank::AbstractItemBank;
    zero_symmetric,
    items,
    thresh
) -> Tuple{Any, Any}

Given an item bank, this function returns the domain of the item bank, i.e. the range (lo, hi) which includes for each item the range in which the the item response function is changing.

Base.eachindex — Method

eachindex(item_bank::AbstractItemBank) -> Base.OneTo

Returns an AbstractUnitRange of item indices for the item bank.

FittedItemBanks.item_params — Method

item_params(
    item_bank::AbstractItemBank,
    idx
) -> NamedTuple{(:discrimination,), <:Tuple{Any}}

Returns the raw parameters for the item at idx as a named tuple. This may return nothing for some item banks. This is debugging/informational use only. Use (ItemResponse)[@ref] for actual item response functions.

ItemResponse methods

FittedItemBanks.resp — Function

resp(ir::ItemResponse, θ) -> Float64  # For BooleanResponse item banks only
resp(ir::ItemResponse, outcome, θ) -> Float64

Return the value of the item response outcome function for the item response ir, the outcome outcome and the ability values θ. For BooleanResponse item banks, outcome can be omitted in which case the outcome is assumed to be true.

FittedItemBanks.resp_vec — Function

resp_vec(ir::ItemResponse, θ) -> AbstractVector{Float64}

Return the vector value of the item response function for the item response ir, the outcome outcome and the ability values θ.

The outcome at each index corresponds with the indices returned by the responses function.

FittedItemBanks.responses — Function

responses(ir::ItemResponse) -> Any

Returns an AbstractVector of possible outcomes for a given (ItemResponse)[@ref].