Non-parametric item banks

Non-parametric IRT models

Sampled and smoothed item banks

FittedItemBanks.DichotomousPointsItemBank — Type

struct DichotomousPointsItemBank <: AbstractItemBank

xs::Vector{Float64}

ys::Matrix{Float64}

An item bank where all items have IRFs computed at a fixed grid across the latent/ability dimension specified as xs. The responses are stored in ys. In most cases this item banks will be coupled with a Smoother and wrapped in a DichotomousSmoothedItemBank.

FittedItemBanks.KernelSmoother — Type

struct KernelSmoother <: Smoother

kernel::Function

bandwidths::Vector{Float64}

A smoother that uses a kernel to smooth the IRF. The bandwidths field stores the kernel bandwidth for each item.

FittedItemBanks.NearestNeighborSmoother — Type

struct NearestNeighborSmoother <: Smoother

Nearest neighbor/staircase smoother.

FittedItemBanks.PointsItemBank — Type

struct PointsItemBank <: AbstractItemBank

xs::Vector{Float64}

ys::ArraysOfArrays.VectorOfArrays{Float64, 2, M, VT} where {M, VT<:AbstractVector{Float64}}

An item bank where all items have IRFs computed at a fixed grid across the latent/ability dimension specified as xs. The responses per-category are stored in ys. In most cases this item banks will be coupled with a Smoother and wrapped in a SmoothedItemBank.

FittedItemBanks.Smoother — Type

abstract type Smoother

FittedItemBanks.gauss_kern — Method

gauss_kern(u)

A guassian kernel for use with KernelSmoother

FittedItemBanks.quad_kern — Method

quad_kern(u)

A quadratic kernel for use with KernelSmoother

FittedItemBanks.uni_kern — Method

uni_kern(u)

A uniform kernel for use with KernelSmoother