Parametric item banks

Parametric IRT models

High-level item banks

The high-level item banks provide shortcuts for common IRT parameterisations.

FittedItemBanks.ItemBank2PL — Method

Convenience function to construct an item bank of the standard 2-parameter logistic single-dimensional IRT model.

FittedItemBanks.ItemBank3PL — Method

Convenience function to construct an item bank of the standard 3-parameter logistic single-dimensional IRT model.

FittedItemBanks.ItemBank4PL — Method

Convenience function to construct an item bank of the standard 4-parameter logistic single-dimensional IRT model.

FittedItemBanks.ItemBankMirt2PL — Method

Convenience function to construct an item bank of the standard 2-parameter logistic MIRT model.

FittedItemBanks.ItemBankMirt3PL — Method

Convenience function to construct an item bank of the standard 3-parameter logistic MIRT model.

FittedItemBanks.ItemBankMirt4PL — Method

Convenience function to construct an item bank of the standard 4-parameter logistic MIRT model.

Composable item banks

The composable item banks allow flexible specification of item banks by combining different blocks to build a variety of model parameterisations.

FittedItemBanks.CdfMirtItemBank — Type

This item bank corresponds to the most commonly found version of MIRT in the literature. Its items feature multidimensional discriminations and its learners multidimensional abilities, but item difficulties are single-dimensional.

FittedItemBanks.NominalItemBank — Type

This item bank implements the nominal model. The Graded Partial Credit Model (GPCM) is implemented in terms of this one. See:

A Generalized Partial Credit Model: Application of an EM Algorithm Muraki, E., (1992). Applied Psychological Measurement 10.1177/014662169201600206

And/or

A Generalized Partial Credit Model Muraki, E. (1997). In Handbook of Modern Item Response Theory. Springer, New York, NY. https://doi.org/10.1007/978-1-4757-2691-6_9

Currently, this item bank only supports the normal scaled logistic as the characteristic/transfer function.