FittedItemBanks.jl

This module provides abstract and concrete item banks, which store information about items and their parameters such as difficulty, most typically resulting from fitting an Item-Response Theory (IRT) model.

FittedItemBanks.FittedItemBanks — Module

This module provides abstract and concrete item banks, which store information about items and their parameters such as difficulty, most typically resulting from fitting an Item-Response Theory (IRT) model.

Contents

Generic interface

Parametric item banks

Non-parametric item banks

Index

FittedItemBanks.FittedItemBanks

FittedItemBanks.AbstractItemBank

FittedItemBanks.BSplineItemBank

FittedItemBanks.BooleanResponse

FittedItemBanks.CdfMirtItemBank

FittedItemBanks.ContinuousDomain

FittedItemBanks.DichotomousPointsItemBank

FittedItemBanks.DichotomousSmoothedItemBank

FittedItemBanks.DiscreteDomain

FittedItemBanks.DiscreteIndexableDomain

FittedItemBanks.DiscreteIterableDomain

FittedItemBanks.DomainType

FittedItemBanks.ItemResponse

FittedItemBanks.KernelSmoother

FittedItemBanks.MonopolyItemBank

FittedItemBanks.MultiGridDichotomousPointsItemBank

FittedItemBanks.MultinomialResponse

FittedItemBanks.NearestNeighborSmoother

FittedItemBanks.NominalItemBank

FittedItemBanks.OneDimContinuousDomain

FittedItemBanks.ResponseType

FittedItemBanks.SlopeInterceptMirtItemBank

FittedItemBanks.SlopeInterceptTransferItemBank

FittedItemBanks.Smoother

FittedItemBanks.TransferItemBank

FittedItemBanks.VectorContinuousDomain

Base.eachindex

Base.length

FittedItemBanks.GPCMItemBank

FittedItemBanks.ItemBank2PL

FittedItemBanks.ItemBank3PL

FittedItemBanks.ItemBank4PL

FittedItemBanks.ItemBankMirt2PL

FittedItemBanks.ItemBankMirt3PL

FittedItemBanks.ItemBankMirt4PL

FittedItemBanks.gauss_kern

FittedItemBanks.gridify

FittedItemBanks.item_bank_domain

FittedItemBanks.item_params

FittedItemBanks.quad_kern

FittedItemBanks.resp

FittedItemBanks.resp_vec

FittedItemBanks.responses

FittedItemBanks.subset

FittedItemBanks.subset_view

FittedItemBanks.uni_kern