Generic interface

This page details the operations which should be supported by different ItemResponse elements, as well as traits for categorisation which can be used to dispatch to different operations.

FittedItemBanks.FittedItemBanks — Module

This module provides abstract and concrete item banks, which store information about items and their parameters such as difficulty, most typically resulting from fitting an Item-Response Theory (IRT) model.

FittedItemBanks.BooleanResponse — Type

struct BooleanResponse <: ResponseType

A boolean/dichotomous response.

FittedItemBanks.ContinuousDomain — Type

abstract type ContinuousDomain <: DomainType

A continuous domain.

FittedItemBanks.DiscreteDomain — Type

abstract type DiscreteDomain <: DomainType

A discrete domain. Typically this is a sampled version of a continuous domain item bank.

Item response functions with discrete domains tend to support less operations than those with continuous domains.

FittedItemBanks.DiscreteIndexableDomain — Type

struct DiscreteIndexableDomain <: DiscreteDomain

An discrete domain which is efficiently indexable and iterable.

FittedItemBanks.DiscreteIterableDomain — Type

struct DiscreteIterableDomain <: DiscreteDomain

An discrete domain which is only efficiently iterable.

FittedItemBanks.DomainType — Type

abstract type DomainType

Domain type for a item banks' item response function.

FittedItemBanks.ItemResponse — Type

struct ItemResponse{ItemBankT<:AbstractItemBank}

item_bank::AbstractItemBank

index::Int64

An item response.

FittedItemBanks.MultinomialResponse — Type

struct MultinomialResponse <: ResponseType

A multinomial response, including ordinal responses.

FittedItemBanks.OneDimContinuousDomain — Type

struct OneDimContinuousDomain <: ContinuousDomain

A continuous domain that is scalar valued.

FittedItemBanks.ResponseType — Type

abstract type ResponseType

A response type for an item bank.

FittedItemBanks.VectorContinuousDomain — Type

struct VectorContinuousDomain <: ContinuousDomain

A continuous domain that is vector valued.

FittedItemBanks._search — Method

Binary search for the point x where f(x) = target += precis given f is assumed as monotonically increasing.

FittedItemBanks.item_bank_domain — Method

Given an item bank, this function returns the domain of the item bank, i.e. the range (lo, hi) which includes for each item the range in which the the item response function is changing.