Parametric item banks

Parametric IRT models

High-level item banks

The high-level item banks provide shortcuts for common IRT parameterisations.

FittedItemBanks.ItemBank2PL — Method

Convenience function to construct an item bank of the standard 2-parameter logistic single-dimensional IRT model.

FittedItemBanks.ItemBank3PL — Method

Convenience function to construct an item bank of the standard 3-parameter logistic single-dimensional IRT model.

FittedItemBanks.ItemBank4PL — Method

Convenience function to construct an item bank of the standard 4-parameter logistic single-dimensional IRT model.

FittedItemBanks.ItemBankMirt2PL — Method

Convenience function to construct an item bank of the standard 2-parameter logistic MIRT model.

FittedItemBanks.ItemBankMirt3PL — Method

Convenience function to construct an item bank of the standard 3-parameter logistic MIRT model.

FittedItemBanks.ItemBankMirt4PL — Method

Convenience function to construct an item bank of the standard 4-parameter logistic MIRT model.

Composable item banks

The composable item banks allow flexible specification of item banks by combining different blocks to build a variety of model parameterisations.

FittedItemBanks.CdfMirtItemBank — Type

This item bank corresponds to the most commonly found version of MIRT in the literature. Its items feature multidimensional discriminations and its learners multidimensional abilities, but item difficulties are single-dimensional.

FittedItemBanks.NominalItemBank — Type

struct NominalItemBank{RankStorageT<:(AbstractVector{<:AbstractArray{<:Real}}), CategoryStorageT<:(AbstractVector{<:AbstractArray{Float64}})} <: AbstractItemBank

This item bank implements the nominal model. The Graded Partial Credit Model (GPCM) is implemented in terms of this one.

Currently, this item bank only supports the normal scaled logistic as the characteristic/transfer function.

References:

A Generalized Partial Credit Model: Application of an EM Algorithm, Muraki, E., (1992). Applied Psychological Measurement.

A Generalized Partial Credit Model, Muraki, E. (1997). In Handbook of Modern Item Response Theory.

FittedItemBanks.MonopolyItemBank — Type

struct MonopolyItemBank <: AbstractItemBank

This item bank implements the monotonic polynomial model with dichotomous responses.

\[\mathrm{irf}(\theta|\xi,{\bf b})=\xi+b_{1}\theta+b_{2}\theta^{2}+\dots+b_{2k+1}\theta^{2k+1}\]

\[\mathrm{irf}^{\prime}(\theta|\mathbf{a})=a_{0}+a_{1}\theta+a_{2}\theta^{2}+\cdot\cdot\cdot+a_{2k}\theta^{2k}\]

References:

Maximum Marginal Likelihood Estimation of a Monotonic Polynomial Generalized Partial Credit Model with Applications to Multiple Group Analysis, Falk, C.F., Cai, L. (2016). Psychometrika.

FittedItemBanks.BSplineItemBank — Type

struct BSplineItemBank <: AbstractItemBank

This item bank implements the a bank with B-spline based item-responses with dichotomous responses.

References:

Maximum Marginal Likelihood Estimation of a Monotonic Polynomial Generalized Partial Credit Model Winsberg, S., Thissen, D. and Wainer, H. (1984) ETS Research Report Series.